Split out rpm.py and hook up rpm logic with zypper as well
We have a bunch of rpm related logic that's required when using
dnf and zypper so let's split out rpm.py and hook up everything in
both dnf and zypper.
