+++ mkosi/distributions/alma.py
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: LGPL-2.1+
 
 from mkosi.distributions import centos, join_mirror
-from mkosi.installer.dnf import RpmRepository, find_rpm_gpgkey
+from mkosi.installer.rpm import RpmRepository, find_rpm_gpgkey
 from mkosi.state import MkosiState
