+++ mkosi/installer/__init__.py
@@ -4,8 +4,9 @@ import os
 
 from mkosi.config import ConfigFeature
 from mkosi.installer.apt import apt_cmd
-from mkosi.installer.dnf import dnf_cmd, rpm_cmd
+from mkosi.installer.dnf import dnf_cmd
 from mkosi.installer.pacman import pacman_cmd
+from mkosi.installer.rpm import rpm_cmd
 from mkosi.installer.zypper import zypper_cmd
 from mkosi.run import apivfs_cmd
 from mkosi.state import MkosiState
